Move type classes into src/ir/types, adjust namespaces & header guards.
