+++ src/ir/types/primitive_type.h
@@ -1,11 +1,11 @@
-#ifndef SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_PRIMITIVE_TYPE_H_
-#define SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_PRIMITIVE_TYPE_H_
+#ifndef SRC_IR_TYPES_PRIMITIVE_TYPE_H_
+#define SRC_IR_TYPES_PRIMITIVE_TYPE_H_
 
 #include "absl/container/flat_hash_set.h"
 #include "src/ir/access_path_selectors_set.h"
-#include "src/transform/ir_proto_to_datalog/arcs_manifest_tree/type.h"
+#include "src/ir/types/type.h"
 
-namespace raksha::transform::arcs_manifest_tree {
+namespace raksha::ir::types {
 
 // TODO: Add the various kinds of primitive type.
 // Right now, the most important thing about primitive types is that they
@@ -25,6 +25,6 @@ class PrimitiveType : public Type {
   }
 };
 
-}  // namespace raksha::transform::arcs_manifest_tree
+}  // namespace raksha::ir::types
 
-#endif  // SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_PRIMITIVE_TYPE_H_
+#endif  // SRC_IR_TYPES_PRIMITIVE_TYPE_H_
