+++ src/ir/types/primitive_type_test.cc
@@ -1,10 +1,10 @@
-#include "src/transform/ir_proto_to_datalog/arcs_manifest_tree/primitive_type.h"
+#include "src/ir/types/primitive_type.h"
 
 #include "src/common/testing/gtest.h"
 #include "src/ir/access_path_selectors.h"
 #include "src/ir/access_path_selectors_set.h"
 
-namespace raksha::transform::arcs_manifest_tree {
+namespace raksha::ir::types {
 
 namespace ir = raksha::ir;
 
@@ -18,4 +18,4 @@ TEST(TestGetAccessPaths, TestGetAccessPaths) {
   ASSERT_EQ(*access_path_selectors_set.begin(), ir::AccessPathSelectors());
 }
 
-}  // namespace raksha::transform::ir_proto_to_datalog
+}  // namespace raksha::ir::types
