+++ src/ir/types/schema.h
@@ -1,14 +1,14 @@
-#ifndef SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_SCHEMA_H_
-#define SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_SCHEMA_H_
+#ifndef SRC_IR_TYPES_SCHEMA_H_
+#define SRC_IR_TYPES_SCHEMA_H_
 
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "absl/types/optional.h"
 #include "src/ir/access_path_selectors_set.h"
-#include "src/transform/ir_proto_to_datalog/arcs_manifest_tree/type.h"
+#include "src/ir/types/type.h"
 #include "third_party/arcs/proto/manifest.pb.h"
 
-namespace raksha::transform::arcs_manifest_tree {
+namespace raksha::ir::types {
 
 class Schema {
  public:
@@ -24,6 +24,6 @@ class Schema {
   absl::flat_hash_map<std::string, std::unique_ptr<Type>> fields_;
 };
 
-}  // namespace raksha::transform::arcs_manifest_tree
+}  // namespace raksha::ir::types
 
-#endif  // SRC_TRANSFORM_IR_PROTO_TO_DATALOG_ARCS_MANIFEST_TREE_SCHEMA_H_
+#endif  // SRC_IR_TYPES_SCHEMA_H_
